action_atlas/sae_backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
ca-certificates=2024.12.31=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cuda-cudart=12.4.127=h99ab3db_0
cuda-cudart_linux-64=12.4.127=hd681fbe_0
cuda-nvrtc=12.4.127=h99ab3db_1
cuda-version=12.4=hbda6634_3
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pypi_0
faiss-gpu=1.9.0=py3.9_ha2e4562_0_cuda12.1.1
flask=3.1.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
itsdangerous=2.2.0=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libcublas=12.4.5.8=h99ab3db_1
libfaiss=1.9.0=h4818125_0_cuda12.1.1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
openai=1.61.1=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.2=py39h06a4308_0
pip=25.0=py39h06a4308_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
python=3.9.21=he870216_1
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
setuptools=75.8.0=py39h06a4308_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h39e8969_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
